+++ test/diffOptions.test.ts
@@ -44,7 +44,7 @@ describe('test Options', () => {
   describe('test OrderBy ASC Option', () => {
     const sqlAlways = new Builder(db, { addAscAfterOrderByItem: 'always' })
     const sqlNever = new Builder(db, { addAscAfterOrderByItem: 'never' })
-    const sqlWhenSpecified = new Builder(db, { addAscAfterOrderByItem: 'when specified' })
+    const sqlWhenMentioned = new Builder(db, { addAscAfterOrderByItem: 'when mentioned' })
     const sqlDefault = new Builder(db)
     it('Produces [SELECT col1 FROM testTable ORDER BY col1 ASC;] option(always)', () => {
       const actual = sqlAlways
@@ -76,8 +76,8 @@ describe('test Options', () => {
       expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1;')
     })
 
-    it('Produces [SELECT col1 FROM testTable ORDER BY col1 ASC;] option(when specified)', () => {
-      const actual = sqlWhenSpecified
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1 ASC;] option(when mentioned)', () => {
+      const actual = sqlWhenMentioned
         .select(column1)
         .from(table)
         .orderBy(column1.asc)
@@ -86,8 +86,8 @@ describe('test Options', () => {
       expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1 ASC;')
     })
 
-    it('Produces [SELECT col1 FROM testTable ORDER BY col1;] option(when specified)', () => {
-      const actual = sqlWhenSpecified
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1;] option(when mentioned)', () => {
+      const actual = sqlWhenMentioned
         .select(column1)
         .from(table)
         .orderBy(column1)
@@ -114,6 +114,82 @@ describe('test Options', () => {
         .getSQL()
 
       expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1 ASC;')
+    })
+  })
+
+  describe('test OrderBy NULLS LAST Option', () => {
+    const sqlAlways = new Builder(db, { addNullsLastAfterOrderByItem: 'always' })
+    const sqlNever = new Builder(db, { addNullsLastAfterOrderByItem: 'never' })
+    const sqlWhenMentioned = new Builder(db, { addNullsLastAfterOrderByItem: 'when mentioned' })
+    const sqlDefault = new Builder(db)
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1 NULLS LAST;] option(always)', () => {
+      const actual = sqlAlways
+        .select(column1)
+        .from(table)
+        .orderBy(column1)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1 NULLS LAST;')
+    })
+
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1;] option(never)', () => {
+      const actual = sqlNever
+        .select(column1)
+        .from(table)
+        .orderBy(column1)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1;')
+    })
+
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1;] option(never) even nulls last mentioned', () => {
+      const actual = sqlNever
+        .select(column1)
+        .from(table)
+        .orderBy(column1.nullsLast)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1;')
+    })
+
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1 ASC;] option(when mentioned)', () => {
+      const actual = sqlWhenMentioned
+        .select(column1)
+        .from(table)
+        .orderBy(column1.asc)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1 ASC;')
+    })
+
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1;] option(when mentioned)', () => {
+      const actual = sqlWhenMentioned
+        .select(column1)
+        .from(table)
+        .orderBy(column1)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1;')
+    })
+
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1;] option(Default)', () => {
+      const actual = sqlDefault
+        .select(column1)
+        .from(table)
+        .orderBy(column1)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1;')
+    })
+
+    it('Produces [SELECT col1 FROM testTable ORDER BY col1 NULLS LAST;] option(Default)', () => {
+      const actual = sqlDefault
+        .select(column1)
+        .from(table)
+        .orderBy(column1.nullsLast)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT col1 FROM testTable ORDER BY col1 NULLS LAST;')
     })
   })
 })